+++ R/taylor-package.R
@@ -23,5 +23,7 @@ NULL
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
 #' @importFrom tibble tibble
+#' @importFrom methods setOldClass
+#' @importFrom lifecycle deprecated is_present deprecate_warn deprecate_soft
 ## usethis namespace: end
 NULL